Split desk provider and reuse menu controls
