+++ apps/ui/src/ui-desks/controls/style.module.css
@@ -1,63 +1,35 @@
-.control {
-	position: relative;
-}
-
 .button {
-	display: inline-flex;
+	--wp-ui-button-height: 36px;
+	--wp-ui-button-background-color: transparent;
+	--wp-ui-button-background-color-active: rgba( 15, 23, 42, 0.07 );
+	--wp-ui-button-border-color: transparent;
+	--wp-ui-button-border-color-active: transparent;
+	--wp-ui-button-foreground-color: #14171a;
+	--wp-ui-button-foreground-color-active: #14171a;
+
 	width: 36px;
-	height: 36px;
-	align-items: center;
-	justify-content: center;
-	padding: 0;
-	border: 0;
+	min-width: 36px;
 	border-radius: 12px;
-	background: transparent;
-	color: #14171a;
-	cursor: pointer;
-	transition:
-		background 140ms ease,
-		color 140ms ease;
 }
 
-.button:hover,
-.button[data-active='true'] {
-	background: rgba( 15, 23, 42, 0.07 );
+.button[aria-pressed='true'] {
+	--wp-ui-button-background-color: rgba( 15, 23, 42, 0.07 );
+	--wp-ui-button-background-color-active: rgba( 15, 23, 42, 0.07 );
+	--wp-ui-button-foreground-color: #14171a;
+	--wp-ui-button-foreground-color-active: #14171a;
 }
 
-.popover {
-	position: absolute;
-	bottom: calc( 100% + 18px );
-	left: 50%;
-	display: flex;
-	align-items: center;
+.swatchMenu {
+	min-width: auto;
+	flex-direction: row;
 	gap: 6px;
 	padding: 6px;
-	border: 1px solid rgba( 255, 255, 255, 0.7 );
 	border-radius: 18px;
-	background: rgba( 255, 255, 255, 0.86 );
-	box-shadow:
-		0 1px 2px rgba( 15, 23, 42, 0.06 ),
-		0 18px 44px rgba( 15, 23, 42, 0.16 );
-	transform: translateX( -50% );
-	transform-origin: bottom center;
-	-webkit-backdrop-filter: saturate( 180% ) blur( 28px );
-	backdrop-filter: saturate( 180% ) blur( 28px );
-	animation: popoverRise 220ms cubic-bezier( 0.32, 0.72, 0.24, 1 );
 }
 
-@keyframes popoverRise {
-	from {
-		opacity: 0;
-		transform: translateX( -50% ) translateY( 4px ) scale( 0.99 );
-	}
-
-	to {
-		opacity: 1;
-		transform: translateX( -50% ) translateY( 0 ) scale( 1 );
-	}
-}
-
-.swatch {
+.swatch,
+.swatch[data-highlighted],
+.swatch:hover {
 	width: 30px;
 	height: 30px;
 	padding: 0;
@@ -65,12 +37,13 @@
 	border-radius: 10px;
 	background: var( --control-swatch-color );
 	box-shadow: 0 6px 18px rgba( 15, 23, 42, 0.1 );
-	cursor: pointer;
+	color: transparent;
 	transition:
 		box-shadow 160ms ease,
 		transform 160ms ease;
 }
 
+.swatch[data-highlighted],
 .swatch:hover {
 	box-shadow: 0 8px 22px rgba( 15, 23, 42, 0.14 );
 	transform: translateY( -1px );
